requirements.txt
# Contains all libraries that are needed for this project

# Data manipulation and analysis
pandas
numpy

# Visualization
matplotlib
seaborn
plotly

# Machine learning
scikit-learn
# xgboost
# lightgbm
# catboost

# Deep learning
# tensorflow
# keras
# torch
# torchvision

# Data processing
# scipy

# Jupyter support
jupyter
notebook

# Model evaluation
# statsmodels

# Utilities
# tqdm

# Optional: Data connectors
# sqlalchemy
# openpyxl

# --- App dependencies for Advanced/Week 3 Streamlit Space ---
streamlit>=1.33.0
# Official Tavily Python client
tavily-python>=0.3.5
# For running on Spaces (optional but common)
python-dotenv>=1.0.1
